ITkannadigaru-HelmCharts-beginer-to-advance-main/Python-Custom-app/requirements.txt
streamlit>=1.26.0
Pillow>=10.0.0